Removing an unneeded call of edit_refresh(), and using 'size_t' instead of 'int' for line lengths. Patch by David Lawrence Ramsey.
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@4895 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/text.c
@@ -420,7 +420,7 @@ void do_undo(void)
     bool gotolinecolumn = FALSE;
     undo *u = openfile->current_undo;
     filestruct *t = 0;
-    int len = 0;
+    size_t len = 0;
     char *undidmsg, *data;
     filestruct *oldcutbuffer = cutbuffer, *oldcutbottom = cutbottom;
 
@@ -555,7 +555,7 @@ void do_redo(void)
 {
     bool gotolinecolumn = FALSE;
     undo *u = openfile->undotop;
-    int len = 0;
+    size_t len = 0;
     char *undidmsg, *data;
 
     for (; u != NULL && u->next != openfile->current_undo; u = u->next)
